Rework how the horizontal width calculations are embedded in the form components to be consistent across the multiple components. Fix a bug where the control block was only building the correct class names when an array of responsive sizes were used rather than a space delimited string.
