+++ resources/views/form/control-group/horizontal.blade.php
@@ -2,8 +2,10 @@
     @if (isset($label))
         <label for="{{ $attributes['id'] }}" class="{{ $horizontalLabelWidth() }} col-form-label">{{ $label }}</label>
     @endif
-    <div class="{{ $horizontalWidth() }} input-group">
-        {{ $slot }}
+    <div class="{{ $horizontalWidth() }}">
+        <div class="input-group">
+            {{ $slot }}
+        </div>
     </div>
     @if (isset($help))
         <div class="form-text">{{ $help }}</div>
