+++ src/Components/Form/AbstractFormComponent.php
@@ -5,7 +5,7 @@ namespace StoneHilt\Bootstrap\Components\Form;
 use Illuminate\Validation\Rule;
 use Illuminate\View\ComponentAttributeBag;
 use StoneHilt\Bootstrap\Components\Base;
-use StoneHilt\Bootstrap\Components\Traits\PrefixNames;
+use StoneHilt\Bootstrap\Components\Traits\HasHorizontalLayoutWithLabel;
 use StoneHilt\Bootstrap\Components\Traits\ResponsiveSizes;
 
 /**
@@ -15,7 +15,7 @@ use StoneHilt\Bootstrap\Components\Traits\ResponsiveSizes;
  */
 abstract class AbstractFormComponent extends Base
 {
-    use PrefixNames;
+    use HasHorizontalLayoutWithLabel;
     use ResponsiveSizes;
 
     /**
@@ -41,42 +41,6 @@ abstract class AbstractFormComponent extends Base
         public string|array $wrapperClass = 'mb-3',
     ) {
         parent::__construct();
-    }
-
-    /**
-     * @return string
-     */
-    public function horizontalLabelWidth(): string
-    {
-        $labelWidths = [];
-        $inputWidths = is_string($this->horizontalWidth) ? explode(' ', $this->horizontalWidth) : $this->horizontalWidth;
-
-        // col-xx-0 does not exist and effectively the same as col-xx-12
-        foreach ($inputWidths as $width) {
-            if (is_numeric($width)) {
-                $labelWidths[] = sprintf(
-                    'col-%d',
-                    (12 - $width) ?: 12
-                );
-            } else {
-                [$responsiveType, $columnWidth] = explode('-', $width);
-                $labelWidths[] = sprintf(
-                    'col-%s-%d',
-                    $responsiveType,
-                    (12 - $columnWidth) ?: 12
-                );
-            }
-        }
-
-        return implode(' ', $labelWidths);
-    }
-
-    /**
-     * @return string
-     */
-    public function horizontalWidth(): string
-    {
-        return implode(' ', $this->prefixNames($this->horizontalWidth, 'col'));
     }
 
     /**
