+++ tests/Feature/Components/Form/ControlGroupTest.php
@@ -140,11 +140,13 @@ class ControlGroupTest extends FeatureTestCase
                     'buttonContent' => $btnContent,
                 ],
                 'expects' => [
-                    '<div class="mb-3">',
+                    '<div class="row mb-3">',
                     sprintf('<label for="%s" class="col-sm-2 col-form-label">%s</label>', $id, $label),
-                    '<div class="col-sm-10 input-group">',
+                    '<div class="col-sm-10">',
+                    '<div class="input-group">',
                     sprintf('<input class="form-control" id="%s" name="%s" type="text">', $id, $name),
                     sprintf('<button class="btn" type="button">%s</button>', $btnContent),
+                    '</div>',
                     '</div>',
                     '</div>',
                 ],
